// src/Model/iTechRestApi/iTechControlColumnEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechControlColumnEnum {
    reference = 0,
    boolean = 1,
    byte = 2,
    short = 3,
    int = 4,
    long = 5,
    float = 6,
    double = 7,
    dateTime = 8,
    string = 9,
    custom = 10,
    byteArray = 11,
    type = 12,
    securityObject = 13,
    user = 14,
    identifier = 15,
    path = 16,
    userGroup = 17
}